+++ src/components/hstack/index.tsx
@@ -1,7 +1,7 @@
 import { PropsWithChildren, useContext } from "react";
 import { View, StyleProp, ViewStyle } from "react-native"
 import { styles } from "./styles"
-import { ContextTheme } from "../provider";
+import { ContextTheme } from "../../provider";
 
 interface VStackProps {
   sx?: StyleProp<ViewStyle>