+++ src/components/vstack/styles.ts
@@ -1,9 +1,9 @@
 import { StyleSheet } from 'react-native';
-import { ITheme } from './../provider';
+import { ITheme } from '../../provider';
 
 export const styles = (config: ITheme) => StyleSheet.create({
   container: {
-    flexDirection: 'row',
+    flexDirection: 'column',
     width: config.space.full
   }
 });